% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hFDR.R
\name{plot.hFDR}
\alias{plot.hFDR}
\title{Produces a plot of the estimated FDR}
\usage{
\method{plot}{hFDR}(
  hFDR.obj,
  cv.obj = NULL,
  sign.lambda = -1,
  log_lambda = T,
  log_cv = F,
  show_legend = !is.null(cv.obj)
)
}
\arguments{
\item{hFDR.obj}{an \code{hFDR} object generated by \code{hFDR::hFDR}.}

\item{cv.obj}{a \code{cv.obj} object generated by \code{cv.model} or
\code{cv.gaussgraph}; or a \code{cv.glmnet} object generated by
\code{glmnet::cv.glmnet}.}

\item{sign.lambda}{either plot against lambda or its negative if
sign.lambda=-1 (default).}

\item{log_lambda}{either plot against log(lambda) (default) or lambda.}

\item{log_cv}{either plot cv.obj in log scale or not (default).}

\item{show_legend}{show legend if True.}
}
\description{
Produces a plot of the estimated FDR
}
\examples{
p <- 100; n <- 300; k <- 15
X <- matrix(rnorm(n*p), n)
nonzero <- sample(p, k)
beta <- 1 * (1:p \%in\% nonzero)
y <- X \%*\% beta + rnorm(n)

n_lambda <- 40
lambda_max <- max(abs(t(scale(X, TRUE, FALSE)) \%*\% scale(y, TRUE, FALSE))) / n
sigma <- sqrt(sum(lm(y ~ X)$residuals^2) / (n-p-1))
lambda_min <- sigma / sqrt(n) / 10
lambda <- lambda_max * (lambda_min/lambda_max)^((0:n_lambda)/n_lambda)

# simple example for Lasso selection. See more examples in the vignettes.
glmnet.cv <- glmnet::cv.glmnet(X, y, alpha = 1, nfolds = 10,
                       intercept = TRUE, standardize = TRUE, family = "gaussian")

hFDR.res <- hFDR(X, y, model = "gausslinear", select = "lasso",
                 lambda = lambda, psi = "pval")

plot(hFDR.res, glmnet.cv)

}
