% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.est.R
\name{precision.est.glasso}
\alias{precision.est.glasso}
\title{Estimate the precision matrix by graphical lasso}
\usage{
precision.est.glasso(X, lambda)
}
\arguments{
\item{X}{n-by-p matrix of multivariate Gaussian variables.}

\item{lambda}{regularity parameter sequence.}
}
\value{
the estimated precision matrix.
}
\description{
Estimate the precision matrix by graphical lasso
}
