% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fit.model.R
\name{pred_fit.mle.gausslinear}
\alias{pred_fit.mle.gausslinear}
\title{Model fitting and predition}
\usage{
pred_fit.mle.gausslinear(X.new, X, y)
}
\arguments{
\item{X.new}{matrix of new values for X at which predictions are to be made.}

\item{X}{matrix of explanatory variables for fitting the MLE model.}

\item{y}{response vector for fitting the MLE model.}
}
\value{
a vector of predicted values of length NROW(X.new)
}
\description{
Fit a MLE model and predict the response value in Gaussian linear model (OLS)
}
