% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fit.model.R
\name{pred_fit.model}
\alias{pred_fit.model}
\title{Model fitting and predition}
\usage{
pred_fit.model(X.new, X, y, select, lambda, pred_fit.mle)
}
\arguments{
\item{X.new}{matrix of new values for X at which predictions are to be made.}

\item{X}{matrix of explanatory variables for fitting the sparse MLE model.}

\item{y}{response vector for fitting the sparse MLE model.}

\item{select}{function that performs variable selection.}

\item{lambda}{regularity parameter sequence.}

\item{pred_fit.mle}{function that predicts the response value by the MLE model.}
}
\value{
A vector of predicted values of length NROW(X.new)
}
\description{
Fit a sparse MLE model and predict the response value
}
