% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler.modelX.gauss}
\alias{sampler.modelX.gauss}
\title{Model-X conditional sampler}
\usage{
sampler.modelX.gauss(X, indices, X.mean, X.cov, sample_size, storage = NULL)
}
\arguments{
\item{X}{n-by-p matrix of explanatory variables.}

\item{indices}{non-empty vector of indices of variables that are to be generated
samples. If \code{indices=1:p}, it generates unconditional samples of all
variables.}

\item{X.mean}{length p mean vector of the explanatory variables.}

\item{X.cov}{p-by-p covariance matrix of explanatory variables.}

\item{sample_size}{the number of samples to be generated.}

\item{storage}{an R environment that holds intermediate results to save
computational time when the function is calling repeatively in
\code{hFDR:::se.modelX}.}
}
\value{
a n-by-length(indices)-by-sample_size array that holds the samples of
variables indexed by indices.
}
\description{
Generates samples of some variables conditioning on the others in the model-X
setting.
}
