% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.logistic}
\alias{select.logistic}
\title{Variable selection by glmnet logistic regression}
\usage{
select.logistic(X, y, lambda)
}
\arguments{
\item{X}{n-by-p matrix of explanatory variables.}

\item{y}{binary response vector of length n.}

\item{lambda}{regularity parameter sequence.}
}
\value{
a logical matrix of size p-by-length(lambda). An entry equals True
iff the corresponding variable (row) is selected at the corresponding lambda
(column).
}
\description{
Variable selection by glmnet logistic regression
}
