% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gaussgraph.R
\name{cv.gaussgraph}
\alias{cv.gaussgraph}
\title{Cross-validation in Gaussian graphical model}
\usage{
cv.gaussgraph(
  X,
  lambda,
  precision.est,
  select = NULL,
  type.measure = c("-loglikelihood", "mse"),
  nfold = 10
)
}
\arguments{
\item{X}{n-by-p matrix of variables.}

\item{lambda}{regularity parameter sequence.}

\item{precision.est}{a function that takes arguments (X, lambda). It
estimates the precision matrix at regularity parameters lambda based on the
observation X. The return is the estimated precision matrix of size p-by-p.}

\item{select}{a function that takes arguments (X, lambda) and does
variable pair selection. The return must be a logical matrix of size p(p-1)/2
by length(lambda). Converting a variable pair (i,j) to the column index and
vice versa can be found in hFDR:::pair_to_index and hFDR:::index_to_pair.}

\item{type.measure}{measurement of the error of the estimated precision
matrix based on the test data. If equals "-loglikelihood", it uses minus
log-likelihood, o.w. if equals "mse", it gets predicted values of \code{X[, j]}
based on \code{X[, -j]} from the test dataset and on the estimated precision
matrix from the training set, and use the averaged mean-squared prediction
error over p such prediction problems as the error measurement.}

\item{nfold}{number of folds in cross-validation.}
}
\value{
An object of class \code{cv.obj} that has similar structure as
\code{cv.glmnet}.
 \item{lambda}{the values of lambda used in the fits.}
 \item{cvm}{the mean cross-validated error - a vector of length
 length(lambda).}
 \item{cvsd}{estimate of standard error of cvm.}
 \item{cvup}{upper curve = cvm+cvsd.}
 \item{cvlo}{lower curve = cvm-cvsd.}
 \item{nzero}{number of selected variables at each lambda. NA if augument
 \code{select} is NULL.}
 \item{name}{a text string indicating type of measure (for plotting purposes).}
 \item{lambda.min}{value of lambda that gives minimum cvm.}
 \item{lambda.1se}{largest value of lambda such that error is within 1
 standard error of the minimum.}
 \item{index}{a one column matrix with the indices of lambda.min and
 lambda.1se.}
}
\description{
Cross-validation for a estimating the precision matrix in the Gaussian
graphical model.
}
\examples{
# Please see vignettes "GaussianGraph"

}
