% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hFDR.R
\name{hFDR}
\alias{hFDR}
\title{The FDR estimator}
\usage{
hFDR(
  X,
  y = NULL,
  model = c("gausslinear", "modelX", "gaussgraph"),
  modelX = "auto.gaussian",
  select = c("lasso", "fs"),
  lambda,
  nlambda = length(lambda),
  psi = "pval",
  se = T,
  n_sample.hfdr = 20,
  n_sample.se = 10,
  n_cores = 1
)
}
\description{
Document will be ready soon
}
