% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hFDR.R
\name{plot.hFDR}
\alias{plot.hFDR}
\title{The plot function}
\usage{
\method{plot}{hFDR}(
  hFDR.obj,
  cv.obj,
  sign.tune = -1,
  log_lambda = T,
  log_cv = F,
  show_legend = T
)
}
\description{
Document will be ready soon
}
